+++ base/src/main/java/proguard/analysis/cpa/jvm/witness/JvmMemoryLocation.java
@@ -19,17 +19,37 @@
 package proguard.analysis.cpa.jvm.witness;
 
 import proguard.analysis.cpa.defaults.LatticeAbstractState;
-import proguard.analysis.cpa.defaults.MemoryLocation;
-import proguard.analysis.cpa.jvm.cfa.edges.JvmCfaEdge;
-import proguard.analysis.cpa.jvm.cfa.nodes.JvmCfaNode;
 import proguard.analysis.cpa.jvm.state.JvmAbstractState;
-import proguard.classfile.MethodSignature;
 
 /**
- * A {@link JvmMemoryLocation} is a {@link MemoryLocation} specific to JVM.
- *
- * @author Dmitry Ivanov
+ * A {@link JvmMemoryLocation} points at a specific location in a certain state of the Jvm. For
+ * example a {@link JvmStackLocation} with index 0 indicates the top of the stack.
  */
-public abstract class JvmMemoryLocation
-    extends MemoryLocation<
-        JvmCfaNode, JvmCfaEdge, MethodSignature, LatticeAbstractState, JvmAbstractState> {}
+public abstract class JvmMemoryLocation {
+
+  /**
+   * Given a JVM state, extract its content in the position represented by this {@link
+   * JvmMemoryLocation}.
+   *
+   * @param jvmState The state from which the value is extracted.
+   * @param defaultValue The value returned if it's not possible to extract the value.
+   * @return The value from the JVM abstract state for the memory location represented by this
+   *     object. Or default value if not possible.
+   * @param <T> The type of the states contained in the JVM state. e.g., for taint analysis this
+   *     would be a {@link proguard.analysis.cpa.defaults.SetAbstractState} containing the taints
+   *     and for value analysis a {@link proguard.analysis.cpa.jvm.domain.value.ValueAbstractState}.
+   */
+  public abstract <T extends LatticeAbstractState<T>> T extractValueOrDefault(
+      JvmAbstractState<T> jvmState, T defaultValue);
+
+  // implementations for Object
+
+  @Override
+  public abstract boolean equals(Object obj);
+
+  @Override
+  public abstract int hashCode();
+
+  @Override
+  public abstract String toString();
+}